refactor: extract search tantivy components
