+++ src/search/Cargo.toml
@@ -46,8 +46,10 @@ serde.workspace = true
 serde_json.workspace = true
 sqlparser.workspace = true
 tantivy.workspace = true
+tantivy_utils.workspace = true
 thiserror.workspace = true
 tokio.workspace = true
+tokio-stream.workspace = true
 tonic.workspace = true
 utoipa.workspace = true
 vector-enrichment.workspace = true
