+++ src/search/src/tantivy/cache.rs
