+++ src/search/src/tantivy/pruner.rs
